requirements.txt
python-telegram-bot==20.8
beautifulsoup4
cryptography
requests